+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="Kure.guraburu_kyuen_tool" >
+    package="kure.guraburu_kyuen_tool" >
 
     <application
         android:allowBackup="true"
