+++ app/src/main/java/kure/guraburu_kyuen_tool/MainActivity.kt
@@ -1,10 +1,11 @@
-package Kure.guraburu_kyuen_tool
+package kure.guraburu_kyuen_tool
 
-import Kure.guraburu_kyuen_tool.util.UserInfoViewPagerAdapter
+import kure.guraburu_kyuen_tool.util.UserInfoViewPagerAdapter
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v4.app.FragmentPagerAdapter
 import android.support.v4.view.ViewPager
+import android.util.Log
 import android.view.View
 import android.widget.Button
 
@@ -14,29 +15,37 @@ class MainActivity : AppCompatActivity(), View.OnClickListener {
     private var mPager : ViewPager? = null;
     private var mAdapter: FragmentPagerAdapter? = null;
     private var mCurrentPage = 0;
+    private val TAG = "HogeHoge"
 
     private val mPageButtonId = arrayOf(R.id.reliefButton,R.id.officialTwitterButton,
             R.id.officialPageButton,R.id.wikiButton,R.id.settingButton);
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        try {
+            super.onCreate(savedInstanceState)
+            setContentView(R.layout.activity_main)
 
-        mPager = findViewById(R.id.pager)
+            mPager = findViewById(R.id.pager)
 
-        mAdapter = UserInfoViewPagerAdapter(supportFragmentManager)
-        mPager!!.setAdapter(mAdapter)
-        mCurrentPage = 0
+            mAdapter = UserInfoViewPagerAdapter(supportFragmentManager)
+            mPager!!.setAdapter(mAdapter)
+            mCurrentPage = 0
 
-        // ボタンにリスナーを設定
-        for (item in mPageButtonId) {
-            var button : Button = findViewById(item)
-            button.setOnClickListener(this)
+            // ボタンにリスナーを設定
+            for (item in mPageButtonId) {
+                Log.v(TAG, "ButtonId:$item")
+                var button: Button = findViewById(item)
+                button.setOnClickListener(this)
+            }
+        }
+        catch (e: Throwable) {
+            Log.e(TAG, "$e")
         }
     }
 
     override fun onClick(v: View) {
         // 押下されたボタンによってFragmentを切り替える
+        Log.v(TAG, "onClick:$v.id")
         when (v.id) {
             mPageButtonId[0] -> {
                 mCurrentPage = 0
