+++ app/src/main/java/kure/guraburu_kyuen_tool/OfficialPageFragment.java
@@ -1,21 +1,21 @@
-package Kure.guraburu_kyuen_tool;
+package kure.guraburu_kyuen_tool;
 
-import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 /**
- * 公式ツイッターページ
+ * 公式ホームページ
  *
  * Created by yuuma on 2017/11/11.
  */
 
-public class OfficialTwitterFragment extends Fragment {
+public class OfficialPageFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_officialTwitter, container, false);
+        return inflater.inflate(R.layout.fragment_official_page, container, false);
     }
 }
