+++ app/src/main/res/layout/activity_main.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Kure.guraburu_kyuen_tool.MainActivity">
+    tools:context="kure.guraburu_kyuen_tool.MainActivity"
+    android:orientation="vertical">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -17,24 +18,27 @@
             android:layout_weight="1"
             android:id="@+id/reliefButton"
             android:text="@string/relief"/>
-        <Button
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:id="@+id/wikiButton"
-            android:text="@string/wiki"/>
+
         <Button
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:id="@+id/officialTwitterButton"
-            android:text="@string/officialTwtitter"/>
+            android:text="@string/officialtwtitter"/>
         <Button
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:id="@+id/officialPageButton"
-            android:text="@string/officialPage"/>
+            android:text="@string/officialpage"/>
+
+        <Button
+            android:id="@+id/wikiButton"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/wiki" />
+
         <Button
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -49,4 +53,4 @@
         android:id="@+id/pager"/>
 
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
