+++ app/src/main/res/layout/fragment_relief.xml
@@ -1,16 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:gravity="center_vertical"
+    >
 
     <LinearLayout
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_gravity="center">
+        android:gravity="center">
+
         <TextView
+            android:id="@+id/g"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/relief"/>
+            android:layout_height="match_parent"
+            android:text="@string/relief" />
     </LinearLayout>
 
 </LinearLayout>
